feat: isolate likebutton in its own component
