+++ front-end/src/components/Global/LikeButton/LikeButton.scoped.css
@@ -1,0 +1,21 @@
+.like-button {
+  color: var(--green);
+  background-color: var(--cream);
+  border: 1px solid var(--green);
+  cursor: pointer;
+  border-radius: 50%;
+  width: 87px;
+  height: 87px;
+  position: relative;
+  top: 220px;
+  left: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: background-color 1s ease-out, color 2s ease-out, border 0.4s ease, transform 0.15s linear;
+}
+
+.like-button-active {
+  background-color: var(--green);
+  color: var(--cream);
+}
